[UI][Select] Fix search value in component state
For example if the select is in a popin, we search for a value in a
certain context, we close the popin, open the popin again, the select
keeps the old search value. We don't want that because the context and
the list can have changed, and therefore we have not the focus nor the
select is open so we don't want that filter again.